[Beta v0.07.25]
=VVV==8.PLT LOG 權限傭有者修改
=VVV==2.A521 新序號 SN:
在[通用-工作站]
目前 :[第一個工作站]
	1.[存入時] -> 3.先檢查 所有必填SN -> 2.繼承舊SN
改變 :[第一個工作站]
	1.[檢查時]->
	2.帶出舊SN資料顯示->[存入時]
	3.先繼承 SN ->
	4.刷入新 SN ->
	5.檢查(當繼承SN空+刷入SN空 = [叫]) ->
	6.如果有刷入 以新的SN為主
	7.檢查 所有必填SN
=VVV==9.提供[通用-工作站] 暫停按鍵 停止計算(免登出用)->改成 90M 登出
=VVV==13.移除製令單 不能包含old 被繼承 or 繼承別張製令單
=VVV==15.取消紅色框框及 [舊]產品-SN身分(登記)--->輸入序號後登記後 序號直接
=VVV==16.清掉在foucus [舊]產品-SN身分(登記)-欄位
=VVV==17.重工-工單製令 (繼承"舊"產品SN序號) 還原修正
=VVV==18.資料庫備份 修正

+++ dtriMES/src/main/java/dtri/com/tw/service/ProductionHeaderService.java
@@ -1227,6 +1227,11 @@ public class ProductionHeaderService {
 				// 確認致令
 				if (phs.size() != 0) {
 					ProductionHeader pheader = phs.get(0);
+					// 檢查 移除製令單 不能包含old 被繼承 or 繼承別張製令單
+					if (productionBodyDao.findAllByPbgidAndPbbsnLikeOrPboldsnLike(data.getLong("ph_pb_g_id"), "old", "old").size() > 0) {
+						return check;
+					}
+					// 移除
 					hoursDao.deleteByproductionRecords(pheader.getProductionRecords());
 					productionHeaderDao.deleteByPhidAndSysheader(data.getLong("ph_id"), true);
 					if (data.getLong("ph_pb_g_id") != 1L) {// 不能移除 NO SN
